+++ .gitignore
@@ -5,6 +5,7 @@
 # Xcode
 #
 build/
+Pods/
 *.pbxuser
 !default.pbxuser
 *.mode1v3
@@ -54,5 +55,3 @@ buck-out/
 
 # Bundle artifact
 *.jsbundle
-
-node_modules/
