+++ package.json
@@ -4,12 +4,16 @@
   "private": true,
   "scripts": {
     "start": "node node_modules/react-native/local-cli/cli.js start",
-    "test": "jest"
+    "test": "jest",
+    "ios": "react-native run-ios",
+    "android": "react-native run-android",
+    "rm-node": "rm -rf node_modules && yarn install"
   },
   "dependencies": {
     "react": "16.8.3",
     "react-native": "^0.59.2",
     "react-native-gesture-handler": "^1.1.0",
+    "react-native-maps": "^0.23.0",
     "react-navigation": "^3.3.0",
     "react-redux": "^6.0.1",
     "redux": "^4.0.1"
